Created XLSX convertor and a duplicate copy of the sidebar.
Realised that the XLSX convertor doesn't work, which is likely to be
because the website is blocking the Apps Script user agent (which
indicates that it's a bot).
Therefore, I've created a copy of the sidebar that I'll alter to only
download the PDF ledger (and not do anything else). I'll also modify
the add-on to work in any Apps Script editor, not just Sheets.

Signed-off-by: Christopher Menon <16004217+cmenon12@users.noreply.github.com>