Started converting from Editor Add-On to Workspace Add-On.
As I can't easily convert the ledger to an XLSX file, I'm going to convert this Editor Add-On to a Workspace Add-On that can run in lots more apps. This will require recreating the sidebar using the Card interface (instead of HTML, JS, and CSS). I'm also renaming the project to reflect that it can only download the ledger.

Signed-off-by: Christopher Menon <16004217+cmenon12@users.noreply.github.com>