Improved storing user properties and made Drive-only.
Made the DOWNLOAD and AUTHORISE buttons filled.
getLoginSection() checks if each of the saved values are undefined, and
substitutes in an empty string if they are.
Because the non-Drive version required a URL from Drive anyway (which
the user would have to visit Drive to retrieve), it makes more sense to
make this a Drive-only add-on. The Drive version provides a much
smoother user experience too.
Renamed processSidebarForm() to processDriveSidebarForm().

Signed-off-by: Christopher Menon <16004217+cmenon12@users.noreply.github.com>