Added manually saving the email to the Sent folder in the mailbox using IMAP.
The config has been modified to reflect separate SMTP and IMAP
configurations, and to add an option whether or not to manually save to
the sent folder.
It will now only connect to the mail server using SSL, not TLS.
Also added the Date and Message-ID headers to the message.

Signed-off-by: Christopher Menon <16004217+cmenon12@users.noreply.github.com>