Added more URLs and documentation, recalculated the Santa Run total.
Added the optional index parameter to getEnthuseAmount(), which is the
index of the matches to use because for some pages the first match is
the fundraising target, not the total.
The Santa Run total now comes from the sum of all the individual Santa
Run pages, instead of the grand total minus the other totals.
This allows for the fact that the grand total may include other
donations that aren't tied to any other specific page.
A new named range was also created for this other total.

Signed-off-by: Christopher Menon <16004217+cmenon12@users.noreply.github.com>