Created EnthuseFundraisingSource class and reorganised everything.
EnthuseFundraisingSource is a class that represents a single
fundraising source on Enthuse. One source can consist of multiple pages.
A source has an amount, number of donors, and fees, as well as its named
range.
These are all calculated upon instantiation, but can be recalculated if
it's a profile page (consisting of all the sources) by removing other
sources that are incorporated into the profile page.

Signed-off-by: Christopher Menon <16004217+cmenon12@users.noreply.github.com>